file-org: move 'scroll' component to widgets folder
