+++ src/components/widgets/omg-scroll/index.js
